Create a common cluster config for tests
