+++ tests/services/test_icebreaker_service.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import os
 import sys
 from pathlib import Path
 from unittest import mock
@@ -11,35 +10,9 @@ from workflows.transport.offline_transport import OfflineTransport
 
 from cryoemservices.services import icebreaker
 from cryoemservices.util.relion_service_options import RelionServiceOptions
+from tests.test_utils.config import cluster_submission_configuration
 
 output_relion_options = dict(RelionServiceOptions())
-
-
-def cluster_submission_configuration(tmp_path):
-    # Create a config file
-    config_file = tmp_path / "config.yaml"
-    with open(config_file, "w") as cf:
-        cf.write("rabbitmq_credentials: rmq_creds\n")
-        cf.write(f"recipe_directory: {tmp_path}/recipes\n")
-        cf.write("slurm_credentials:\n")
-        cf.write(f"  default: {tmp_path}/slurm_credentials.yaml\n")
-    os.environ["USER"] = "user"
-
-    # Create dummy slurm credentials files
-    with open(tmp_path / "slurm_credentials.yaml", "w") as slurm_creds:
-        slurm_creds.write(
-            "user: user\n"
-            "user_home: /home\n"
-            f"user_token: {tmp_path}/token.txt\n"
-            "required_directories: [directory1, directory2]\n"
-            "partition: partition\n"
-            "partition_preference: preference\n"
-            "cluster: cluster\n"
-            "url: /url/of/slurm/restapi\n"
-            "api_version: v0.0.40\n"
-        )
-    with open(tmp_path / "token.txt", "w") as token:
-        token.write("token_key")
 
 
 @pytest.fixture
